backend/requirements.txt
# FastAPI
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0

# HTTP Client (Graph API)
httpx==0.27.0

# Auth
msal==1.31.0
PyJWT==2.9.0

# Settings
pydantic-settings==2.5.0

# Utils
python-dotenv==1.0.1
